backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.12.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
chromadb>=0.4.18
sentence-transformers>=2.2.2
openai>=0.28.1
httpx>=0.25.2
python-dotenv>=1.0.0
numpy>=1.24.0,<2.0.0
redis>=5.0.1
PyPDF2>=3.0.1
python-docx>=0.8.11
elasticsearch==8.15.0
spacy>=3.7.0
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.3.0
